chore: rename to no-empty-strings
